+++ backup/moodle2/restore_moochat_stepslib.php
@@ -5,7 +5,6 @@
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
-
 /**
  * Define all the restore steps that will be used by the restore_moochat_activity_task
  *
@@ -13,30 +12,24 @@
  * @copyright  2025 Brian A. Pool
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-
 defined('MOODLE_INTERNAL') || die();
-
 class restore_moochat_activity_structure_step extends restore_activity_structure_step {
-
     /**
      * Defines structure of path elements to be processed during the restore
      *
      * @return array of {@link restore_path_element}
      */
     protected function define_structure() {
-
         $paths = [];
         $userinfo = $this->get_setting_value('userinfo');
-
         $paths[] = new restore_path_element('moochat', '/activity/moochat');
         if ($userinfo) {
             $paths[] = new restore_path_element('moochat_usage', '/activity/moochat/usages/usage');
+            $paths[] = new restore_path_element('moochat_conversation', '/activity/moochat/conversations/conversation');
         }
-
         // Return the paths wrapped into standard activity structure.
         return $this->prepare_activity_structure($paths);
     }
-
     /**
      * Process the given restore path element data
      *
@@ -44,19 +37,15 @@ class restore_moochat_activity_structure_step extends restore_activity_structure
      */
     protected function process_moochat($data) {
         global $DB;
-
         $data = (object)$data;
         $data->course = $this->get_courseid();
-
         $data->timecreated = $this->apply_date_offset($data->timecreated);
         $data->timemodified = $this->apply_date_offset($data->timemodified);
-
         // Insert the moochat record.
         $newitemid = $DB->insert_record('moochat', $data);
         // Immediately after inserting "activity" record, call this.
         $this->apply_activity_instance($newitemid);
     }
-
     /**
      * Process the given restore path element data
      *
@@ -64,19 +53,32 @@ class restore_moochat_activity_structure_step extends restore_activity_structure
      */
     protected function process_moochat_usage($data) {
         global $DB;
-
         $data = (object)$data;
         $oldid = $data->id;
-
         $data->moochatid = $this->get_new_parentid('moochat');
         $data->userid = $this->get_mappingid('user', $data->userid);
         $data->firstmessage = $this->apply_date_offset($data->firstmessage);
         $data->lastmessage = $this->apply_date_offset($data->lastmessage);
-
         $newitemid = $DB->insert_record('moochat_usage', $data);
         $this->set_mapping('moochat_usage', $oldid, $newitemid);
     }
-
+    
+    /**
+     * Process the given restore path element data for conversations
+     *
+     * @param array $data parsed element data
+     */
+    protected function process_moochat_conversation($data) {
+        global $DB;
+        $data = (object)$data;
+        $oldid = $data->id;
+        $data->moochatid = $this->get_new_parentid('moochat');
+        $data->userid = $this->get_mappingid('user', $data->userid);
+        $data->timecreated = $this->apply_date_offset($data->timecreated);
+        $newitemid = $DB->insert_record('moochat_conversations', $data);
+        $this->set_mapping('moochat_conversation', $oldid, $newitemid);
+    }
+    
     /**
      * Post-execution actions
      */