+++ db/upgrade.php
@@ -7,7 +7,7 @@
 // (at your option) any later version.
 
 /**
- * Define all the restore steps that will be used by the restore_moochat_activity_task
+ * Upgrade script for mod_moochat
  *
  * @package    mod_moochat
  * @copyright  2025 Brian A. Pool
@@ -16,56 +16,38 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-/**
- * Upgrade code for mod_moochat
- *
- * @param int $oldversion the version we are upgrading from
- * @return bool always true
- */
 function xmldb_moochat_upgrade($oldversion) {
     global $DB;
-
     $dbman = $DB->get_manager();
 
-    // Add display field
-    if ($oldversion < 2025103002) {
-        $table = new xmldb_table('moochat');
-        $field = new xmldb_field('display', XMLDB_TYPE_INTEGER, '2', null, XMLDB_NOTNULL, null, '0', 'introformat');
-        
-        if (!$dbman->field_exists($table, $field)) {
-            $dbman->add_field($table, $field);
-        }
-        
-        upgrade_mod_savepoint(true, 2025103002, 'moochat');
-    }
+    if ($oldversion < 2026021801) {
+        // Define table moochat_conversations to be created.
+        $table = new xmldb_table('moochat_conversations');
 
-    // Add include_section_content and include_hidden_content fields
-    if ($oldversion < 2025103003) {
-        $table = new xmldb_table('moochat');
-        
-        $field1 = new xmldb_field('include_section_content', XMLDB_TYPE_INTEGER, '1', null, XMLDB_NOTNULL, null, '0', 'display');
-        if (!$dbman->field_exists($table, $field1)) {
-            $dbman->add_field($table, $field1);
-        }
-        
-        $field2 = new xmldb_field('include_hidden_content', XMLDB_TYPE_INTEGER, '1', null, XMLDB_NOTNULL, null, '0', 'include_section_content');
-        if (!$dbman->field_exists($table, $field2)) {
-            $dbman->add_field($table, $field2);
-        }
-        
-        upgrade_mod_savepoint(true, 2025103003, 'moochat');
-    }
+        // Adding fields to table moochat_conversations.
+        $table->add_field('id', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, XMLDB_SEQUENCE, null);
+        $table->add_field('moochatid', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, null);
+        $table->add_field('userid', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, null);
+        $table->add_field('role', XMLDB_TYPE_CHAR, '20', null, XMLDB_NOTNULL, null, null);
+        $table->add_field('message', XMLDB_TYPE_TEXT, null, null, XMLDB_NOTNULL, null, null);
+        $table->add_field('timecreated', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, null);
 
-    // Add chatsize field
-    if ($oldversion < 2025103004) {
-        $table = new xmldb_table('moochat');
-        $field = new xmldb_field('chatsize', XMLDB_TYPE_CHAR, '10', null, XMLDB_NOTNULL, null, 'medium', 'include_hidden_content');
-        
-        if (!$dbman->field_exists($table, $field)) {
-            $dbman->add_field($table, $field);
+        // Adding keys to table moochat_conversations.
+        $table->add_key('primary', XMLDB_KEY_PRIMARY, ['id']);
+        $table->add_key('moochatid', XMLDB_KEY_FOREIGN, ['moochatid'], 'moochat', ['id']);
+        $table->add_key('userid', XMLDB_KEY_FOREIGN, ['userid'], 'user', ['id']);
+
+        // Adding indexes to table moochat_conversations.
+        $table->add_index('moochatid-userid', XMLDB_INDEX_NOTUNIQUE, ['moochatid', 'userid']);
+        $table->add_index('timecreated', XMLDB_INDEX_NOTUNIQUE, ['timecreated']);
+
+        // Conditionally launch create table for moochat_conversations.
+        if (!$dbman->table_exists($table)) {
+            $dbman->create_table($table);
         }
-        
-        upgrade_mod_savepoint(true, 2025103004, 'moochat');
+
+        // Moochat savepoint reached.
+        upgrade_mod_savepoint(true, 2026021801, 'moochat');
     }
 
     return true;